0x0F-pointer_functions projects task #2 code update

+++ 0x0F-function_pointers/2-int_index.c
@@ -7,7 +7,7 @@
  * Return: index of first element for which cmp function does not return 0
  * -1 if no element matches or size <= 0
  */
-int int_index(int *array, int size, int (*cmp)(int))
+int int_index(int *array, size_t size, int (*cmp)(int))
 {
 	unsigned int i;
